I added a deck class
I added a deck class and generalize the interleaved reading of the
audio input from "ofxSoundStream" so that the <deck> tag in
"settings.xml" can set how many inputs and turntables are present at
run. Changed names of classes and a "Includes.h" Things I broke OSC
from the turntables…Needs a love.